# nimo-plm/api/openapi.yaml
openapi: 3.1.0
info:
  title: nimo PLM System API
  description: |
    nimo智能眼镜PLM系统API接口规范
    
    ## 认证方式
    所有接口（除/auth/*外）需要在Header中携带JWT Token：
    ```
    Authorization: Bearer <token>
    ```
    
    ## 通用响应格式
    ```json
    {
      "code": 0,
      "message": "success",
      "data": {...}
    }
    ```
    
    ## 错误码
    - 0: 成功
    - 400xx: 参数错误
    - 401xx: 认证错误
    - 403xx: 权限错误
    - 404xx: 资源不存在
    - 500xx: 服务器错误
  version: 1.0.0
  contact:
    name: nimo PLM Team
    email: plm@bitfantasy.com
servers:
  - url: http://localhost:8001
    description: 开发环境
  - url: https://plm-api.nimo.internal
    description: 生产环境
...
paths:
...
  /api/v1/auth/feishu/login:
    get:
      tags: [Auth]
      summary: 飞书登录
      description: 重定向到飞书OAuth授权页面
      operationId: feishuLogin
      parameters:
        - name: redirect_uri
          in: query
          description: 授权后回调地址
          schema:
            type: string
      responses:
        '302':
          description: 重定向到飞书授权页面
  /api/v1/auth/feishu/callback:
    get:
      tags: [Auth]
      summary: 飞书授权回调
      description: 飞书授权完成后的回调接口
      operationId: feishuCallback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: 刷新Token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
              required:
                - refresh_token
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
...
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: 获取当前用户信息
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserDetail'

  # ============================================================
  # 用户管理接口
  # ============================================================
...
  /api/v1/projects:
    get:
      tags: [Projects]
      summary: 获取项目列表
      operationId: listProjects
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: keyword
          in: query
          schema:
            type: string
        - name: product_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [planning, evt, dvt, pvt, mp, completed, cancelled]
        - name: owner_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProjectListResponse'

    post:
      tags: [Projects]
      summary: 创建项目
      operationId: createProject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Project'
  /api/v1/projects/{id}:
    get:
      tags: [Projects]
      summary: 获取项目详情
      operationId: getProject
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProjectDetail'

    put:
      tags: [Projects]
      summary: 更新项目
      operationId: updateProject
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Project'
...
  /api/v1/projects/{projectId}/tasks:
    get:
      tags: [Tasks]
      summary: 获取项目任务列表
      operationId: listProjectTasks
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: phase
          in: query
          schema:
            type: string
            enum: [evt, dvt, pvt, mp]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed, blocked, cancelled]
        - name: assignee_id
          in: query
          schema:
            type: string
        - name: overdue_only
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Task'

    post:
      tags: [Tasks]
      summary: 创建任务
      operationId: createTask
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Task'
  /api/v1/tasks/{id}:
    get:
      tags: [Tasks]
      summary: 获取任务详情
      operationId: getTask
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TaskDetail'

    put:
      tags: [Tasks]
      summary: 更新任务
      operationId: updateTask
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Task'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: 页码
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSizeParam:
      name: page_size
      in: query
      description: 每页数量
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  schemas:
...
    BaseResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer

    # ============================================================
    # 认证相关
    # ============================================================
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                access_token:
                  type: string
                refresh_token:
                  type: string
                expires_in:
                  type: integer
                  description: Token过期时间（秒）
                user:
                  $ref: '#/components/schemas/UserDetail'

    # ============================================================
    # 用户相关
    # ============================================================
    UserSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        avatar_url:
          type: string
    UserDetail:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        name:
          type: string
        email:
          type: string
        mobile:
          type: string
        avatar_url:
          type: string
        department:
          $ref: '#/components/schemas/Department'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        permissions:
          type: array
          items:
            type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
...
    Department:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    Role:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string

    # ============================================================
    # 产品相关
    # ============================================================
    Category:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        parent_id:
          type: string
        children:
          type: array
          items:
            $ref: '#/components/schemas/Category'
    Product:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        category:
          $ref: '#/components/schemas/Category'
        status:
          type: string
          enum: [draft, developing, active, discontinued]
        current_bom_version:
          type: string
        thumbnail_url:
          type: string
        created_by:
          $ref: '#/components/schemas/UserSummary'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ProjectSummary:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        status:
          type: string
        current_phase:
          type: string
        progress:
          type: integer
    Project:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        product:
          $ref: '#/components/schemas/Product'
        status:
          type: string
          enum: [planning, evt, dvt, pvt, mp, completed, cancelled]
        current_phase:
          type: string
        progress:
          type: integer
        owner:
          $ref: '#/components/schemas/UserSummary'
        planned_start:
          type: string
          format: date
        planned_end:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
    ProjectDetail:
      allOf:
        - $ref: '#/components/schemas/Project'
        - type: object
          properties:
            description:
              type: string
            phases:
              type: array
              items:
                $ref: '#/components/schemas/ProjectPhase'
            task_summary:
              type: object
              properties:
                total:
                  type: integer
                completed:
                  type: integer
                in_progress:
                  type: integer
                pending:
                  type: integer
                blocked:
                  type: integer
    ProjectPhase:
      type: object
      properties:
        id:
          type: string
        phase:
          type: string
        name:
          type: string
        status:
          type: string
        planned_start:
          type: string
          format: date
        planned_end:
          type: string
          format: date
        actual_start:
          type: string
          format: date
        actual_end:
          type: string
          format: date
    ProjectListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
        product_id:
          type: string
        description:
          type: string
        template:
          type: string
          enum: [standard, fast_track, custom]
          default: standard
        owner_id:
          type: string
        planned_start:
          type: string
          format: date
        planned_end:
          type: string
          format: date
      required:
        - name
        - product_id
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
        owner_id:
          type: string
        planned_start:
          type: string
          format: date
        planned_end:
          type: string
          format: date
...
    Task:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        project:
          $ref: '#/components/schemas/ProjectSummary'
        phase:
          type: string
        task_type:
          type: string
          enum: [task, milestone, deliverable]
        status:
          type: string
          enum: [pending, in_progress, completed, blocked, cancelled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        assignee:
          $ref: '#/components/schemas/UserSummary'
        progress:
          type: integer
        planned_start:
          type: string
          format: date
        planned_end:
          type: string
          format: date
        due_date:
          type: string
          format: date
        is_overdue:
          type: boolean
    TaskDetail:
      allOf:
        - $ref: '#/components/schemas/Task'
        - type: object
          properties:
            description:
              type: string
            reviewer:
              $ref: '#/components/schemas/UserSummary'
            actual_start:
              type: string
              format: date
            actual_end:
              type: string
              format: date
            estimated_hours:
              type: number
            actual_hours:
              type: number
            parent_task:
              $ref: '#/components/schemas/Task'
            subtasks:
              type: array
              items:
                $ref: '#/components/schemas/Task'
            dependencies:
              type: array
              items:
                $ref: '#/components/schemas/Task'
            comments:
              type: array
              items:
                $ref: '#/components/schemas/TaskComment'
            attachments:
              type: array
              items:
                $ref: '#/components/schemas/Attachment'
    TaskComment:
      type: object
      properties:
        id:
          type: string
        user:
          $ref: '#/components/schemas/UserSummary'
        content:
          type: string
        created_at:
          type: string
          format: date-time
    CreateTaskRequest:
      type: object
      properties:
        name:
          type: string
        phase_id:
          type: string
        parent_task_id:
          type: string
        description:
          type: string
        task_type:
          type: string
          enum: [task, milestone, deliverable]
        priority:
          type: string
          enum: [low, medium, high, critical]
        assignee_id:
          type: string
        reviewer_id:
          type: string
        planned_start:
          type: string
          format: date
        planned_end:
          type: string
          format: date
        estimated_hours:
          type: number
      required:
        - name
    UpdateTaskRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        assignee_id:
          type: string
        progress:
          type: integer
          minimum: 0
          maximum: 100
        planned_start:
          type: string
          format: date
        planned_end:
          type: string
          format: date
        actual_hours:
          type: number

    # ============================================================
    # ECN相关
    # ============================================================
...
    Attachment:
      type: object
      properties:
        id:
          type: string
        file_name:
          type: string
        file_size:
          type: integer
        mime_type:
          type: string
        uploaded_by:
          $ref: '#/components/schemas/UserSummary'
        created_at:
          type: string
          format: date-time
security:
  - bearerAuth: []